requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
